Update to .NET 8
It has much better scaffolding capabilities for SQLite since https://github.com/dotnet/efcore/pull/30816

+++ tests/SqliteTest.Scaffold/Artist.verified.cs
@@ -18,7 +18,7 @@ namespace Scaffold;
 
 public partial class Artist
 {
-    public long ArtistId { get; set; }
+    public int ArtistId { get; set; }
 
     public string? Name { get; set; }
